requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-crontab==0.7.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
mysqlclient==2.2.4
numpy==2.0.1
openai==1.36.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
